requirements.txt
seaborn
matplotlib
numpy
pandas
torch
torchmetrics
torch-dreams
hydra-core
